requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
librosa>=0.9.0
soundfile>=0.10.3
audioread>=2.1.9
scikit-learn>=1.0.0
torch>=2.0.0
torchaudio>=2.0.0
torchcodec>=0.0.1
transformers>=4.37.0
accelerate>=0.20.0
sentencepiece>=0.1.99
protobuf>=3.20.0
tqdm>=4.62.0
h5py>=3.1.0
pyyaml>=6.0
matplotlib>=3.4.0
seaborn>=0.11.0
requests>=2.28.0
einops>=0.6.0
safetensors>=0.3.0
fairlearn==0.13.0
python-Levenshtein
